Added DestinationDistance.kt
Contains 3 Methods, distance(double[] cord1,double[] cord2) is a helper method that calculates the distance between 2 [lat,long] arrays.  totalDistance(double[][] cords) takes in an array of [lat, long] arrays and uses the distance() helper to add up add the distances between points.  estimatedTimeWithDistance(double velocity, double[][] cords) calculates the estimated time given constant velocity with totalDistance() and displays a string with time in "hours:minutes and totalDistance".  All methods are made with the assumption that the array of lat long values are given as nested arrays of the coordinates [[lat1,long1], [lat2,long2], [lat3,long3], etc].  Methods are not used for the application yet.